Moved common build_raw_record method to test helper (#413)
* Moved common build_raw_record method to test helper

* DRY colorectal test classes

* Bump Ruby to 3.0.4 in live, resolves plan.io #29786

* Moved common build_raw_record method to test helper

* DRY colorectal test classes

* PR review comments fix to separate BRCA test helper common method

* Corrected the name for test helper
