+++ test/lib/import/brca/providers/birmingham/birmingham_handler_newformat_test.rb
@@ -300,17 +300,6 @@ class BirminghamHandlerNewformatTest < ActiveSupport::TestCase
 
   private
 
-  def build_raw_record(options = {})
-    default_options = { 'pseudo_id1' => '',
-                        'pseudo_id2' => '',
-                        'encrypted_demog' => '',
-                        'clinical.to_json' => clinical_json,
-                        'encrypted_rawtext_demog' => '',
-                        'rawtext_clinical.to_json' => rawtext_clinical_json }
-
-    Import::Brca::Core::RawRecord.new(default_options.merge!(options))
-  end
-
   def variant_processor_for(record)
     Import::Brca::Providers::Birmingham::VariantProcessor.new(@genotype, record, @logger)
   end
