+++ test/lib/import/brca/providers/salisbury/salisbury_handler_test.rb
@@ -108,21 +108,7 @@ class SalisburyHandlerTest < ActiveSupport::TestCase
 
   private
 
-  # TODO: DRY this method up into a helper
-  def build_raw_record(options = {})
-    default_options = {
-      'pseudo_id1' => '',
-      'pseudo_id2' => '',
-      'encrypted_demog' => '',
-      'clinical.to_json' => clinical_to_json,
-      'encrypted_rawtext_demog' => '',
-      'rawtext_clinical.to_json' => rawtext_to_clinical_to_json
-    }
-
-    Import::Brca::Core::RawRecord.new(default_options.merge!(options))
-  end
-
-  def clinical_to_json
+  def clinical_json
     { sex: '2',
       consultantcode: 'Consultant Code',
       receiveddate: '2017-06-20T00: 00: 00.000+01: 00',
@@ -133,7 +119,7 @@ class SalisburyHandlerTest < ActiveSupport::TestCase
       age: 999 }.to_json
   end
 
-  def rawtext_to_clinical_to_json
+  def rawtext_clinical_json
     { sex: 'Female',
       providercode: 'Provider Name',
       consultantname: 'Consultant Name',
